Add --keep-local functionality
With the change it is now possible to switch the aosp manifest branch
but to keep the sodp local_manifests branch as is. This can be used to
get earlier access to new android security releases.

An overview of the latest security branches from aosp can be found here:
https://source.android.com/setup/start/build-numbers

Having a local_manifests branch ithat is not inline with aosp manifest
branch may cause build-problems or runtime problems. Additionally it is
not a supported configuration by sodp, so failures because of using
different branches must not be reported on the sodp bug_tracker.